backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq==0.11.0
faiss-cpu
numpy>=2.0.0
python-dotenv==1.0.1
httpx==0.27.2
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
cryptography==41.0.7
bcrypt==4.1.1
PyJWT==2.12.0
pdfplumber==0.9.0
openpyxl==3.1.5
python-multipart==0.0.6
reportlab==4.0.9
